refactor(abstract-utxo): move AbstractUtxoCoinWallet to its own file
We want to add more code here

Issue: BTC-1450
